+++ src/main/java/Basic/L3_hw2/BadPupil.java
@@ -3,22 +3,22 @@ package Basic.L3_hw2;
 public class BadPupil extends Pupil {
 
     @Override
-    void read() {
+    protected void read() {
         System.out.println("Плохо читает");
     }
 
     @Override
-    void write() {
+    protected void write() {
         System.out.println("Плохо пишет");
     }
 
     @Override
-    void study() {
+    protected void study() {
         System.out.println("Плохо учится");
     }
 
     @Override
-    void relax() {
+    protected void relax() {
         System.out.println("Плохо отдыхает");
     }
 }
