+++ src/main/java/Basic/L3_hw2/Pupil.java
@@ -2,19 +2,19 @@ package Basic.L3_hw2;
 
 public class Pupil {
 
-    void study() {
+    protected void study() {
         System.out.println("Просто учится");
     }
 
-    void read() {
+    protected void read() {
         System.out.println("Просто читает");
     }
 
-    void write() {
+    protected void write() {
         System.out.println("Просто пишет");
     }
 
-    void relax() {
+    protected void relax() {
         System.out.println("Просто отдыхает");
     }
 }
